Enable useAutoStoreData
We need a way to fake logins, I figure the easiest thing is to store the
data from an invisible field within the login form. When the user enters
anything there'll now be a:

`req.session.data.signedIn` bit of data. this can be accessed from
within the views with data['signedIn'] which we can use to conditionally
set the header.

Sign-out just deletes the key, and hey presto we have login!