Add workflow template to deploy to GOV.UK PaaS
This repo is currently linked to and deployed on GOV.UK's Heroku account
however the team that manages it has moved programmes within GDS and
will soon lose access to the shared Heroku account.

Deploying previews to PaaS instead via Github actions will mean that the
config and secrets for previews is all managed within the repo, which
will be easier to maintain in the future.

This commit adds a template action and manifest for deploying the
prototype to PaaS. This template will be called by separate workflows
that pass the required parameters.

Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet.office.gov.uk>