Add workflow to delete old preview deployments
This is triggered in two ways. Firstly, it runs when a PR is closed
or merged. In that case it looks in PaaS for an app with the name
expected for the branch name, and deletes it.

It also runs on a schedule at 10am every day and deletes any deployed
apps that are older than 30 days. We'll have to update this behaviour
slightly so that it doesn't delete any 'production' prototypes that we
want to keep running for longer than that.