Add action to deploy a long-lived prototype on release
When a new release is created, the action will take the contents of the
`main` branch and deploy it to GOV.UK PaaS as a long lived, stable
prototype for user testing.

Each release prototype gets its own environment named off the release
tag so it won't be overwritten by subsequent deploys.