Fix delete deployment workflow
`cf events` doesn't contain any information for an app running in the
Node.js buildpack, so the scheduled check for old deployments would
always fail to find any. Switch to getting the upload date from the
output of `cf apps` which is always present.

Note - the `date` command works differently on a Mac, but I've tested
these commands in bash in an Ubuntu Docker container and it works as
expected.