Fix step to delete old deployments
This step is meant to filter for all deployments that are older than 30
days and delete any that match the naming convention for branch preview
deploys.

However, the `date` command to parse the date from the output of
`cf app $app` always failed which meant this step was deleting all
branch preview deploys every time it ran.

An example of this failure can be seen at [[1]].

`cf app $app` outputs a date string like `Thu 05 May 16:58:33 BST 2022'`
which isn't a format that `date` can parse. This commit adds an extra
string processing step that switches the position of the day and month
in the string, resulting in something like
`Thu May 05 16:58:33 BST 2022` which `date` can parse.

[1]: https://github.com/alphagov/govuk-accounts-explore-prototype-1/runs/6304209328?check_suite_focus=true