Run delete job at 4 am
The job was scheduled to run every weekday at 10 am UTC, but coupled
with the bug fixed in #6e2e43d this meant that live prototypes were
being deleted during the work day.

If something else goes wrong in the future, it'll be much better for a
prototype to either be running for the whole day or deleted early in
the morning so we have an opportunity to fix it before we start working
with it.